+++ lib/activejob/temporal/temporal_options.rb
@@ -38,6 +38,7 @@ module ActiveJob
         heartbeat_timeout
       ].freeze
 
+      # rubocop:disable Metrics/BlockLength
       class_methods do
         # Define Temporal activity timeout options for this job class.
         #
@@ -104,6 +105,7 @@ module ActiveJob
           end
         end
       end
+      # rubocop:enable Metrics/BlockLength
     end
   end
 end